chore(plugin): rename plugin module file
